requirements.txt
telethon
pandas
openpyxl
python-dotenv
pysocks
